prevented ligne with deblocages linked to subventions where is_versee=true to be deleted by non-admin users

+++ compta/views.py
@@ -201,7 +201,14 @@ def delete_ligne(request, id_ligne):
 		ligne =	LigneCompta.objects.get(id=id_ligne)
 		if not ligne.is_locked or request.user.is_staff:
 			# si l'écriture est locked, seuls les admins peuvent la supprimer
-			ligne.delete()
+			# de même si des déblocages versés ont été fait dessus
+			deblocages_subvention = DeblocageSubvention.objects.filter(ligne_compta=ligne)
+			none_versee = True
+			for deblocage in deblocages_subvention:
+				if deblocage.subvention.is_versee:
+					none_versee = False
+			if none_versee or request.user.is_staff:
+				ligne.delete()
 	return redirect('../')
